added parent class to attractions
