+++ Attractions/__init__.py
@@ -1,3 +1,4 @@
+from .attraction import Attraction
 from .petting_zoo import PettingZoo
 from .snake_pit import SnakePit
 from .wetlands import Wetlands
