+++ Attractions/petting_zoo.py
@@ -1,6 +1,6 @@
-class PettingZoo:
+from .attraction import Attraction
 
-    def __init__(self, name):
-        self.attraction_name = name
-        self.description = "cute and fuzzy critters to cuddle"
-        self.animals = list()
+class PettingZoo(Attraction):
+
+    def __init__(self, name, description):
+        super().__init__(name, description)
