+++ index.py
@@ -16,7 +16,7 @@ from Attractions import Wetlands, PettingZoo, SnakePit
 #     print(f'you can find {animal.name} the {animal.species} in the {crocodile_consortium.attraction_name}')
 
 
-slither_inn = SnakePit("Slither Inn")
+slither_inn = SnakePit("Slither Inn", "we cold-blooded over here")
 jose = Cobra('Jose', 'cobra', 'spaghetti', 666)
 mike = King_snake("Mike", "king Snake", "other snakes", 777)
 
@@ -30,15 +30,3 @@ for animal in slither_inn.animals:
     print(f'-{animal.name} the {animal.species}')
 
 print(slither_inn.last_critter_added)
-
-jeff = Bearded_dragon('Jeff', 'lizard', 'popcorn', 'afternoon', 999)
-
-print(jeff.species)
-print(jeff.shift)
-print(jeff.chip_number)
-print(jeff.feed())
-
-clyde = Turtle('Clyde', 'turtle','steak tartar', 222)
-
-print(clyde.feed())
-clyde.run()
